Remove the scripts I wrote when I was being dumb
Recall that I had the problem of not being able to load NaCl into the context
of my background script.  I thought the problem was some nonsense with the way
"security" works in browser extensions.  There are some contexts where in order
to load a script, you must explicitly specify the SHA-256 or SHA-384 hash of
the script, in base64 notation.  I wrote these scripts for the purpose of
quickly reading a file and spitting out those values, so that I could easily
update the manifest as I was fooling around with the scripts.

It turned out that entire hypothesis was wrong: https://github.com/aeternity/Vanillae/commit/b680336ba464eafe63daafd95002f09e798c09d9

And therefore these scripts are now just junk